refactor(files): changes files structure
