+++ front-end/src/App.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import Sidebar from './componentes/Sidebar';
+import ExibeCavalos from './telas/ExibeCavalos';
+import ExibeProprietarios from './telas/ExibeProprietarios';
+
+const App = () => {
+  return (
+    <Router>
+      <div className="app">
+        <Sidebar />
+        <div className="content">
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/about" element={<About />} />
+          </Routes>
+        </div>
+      </div>
+    </Router>
+  );
+};
+
+export default App;
